+++ api/src/main/java/org/allaymc/api/blockentity/interfaces/BlockEntityDecoratedPot.java
@@ -1,0 +1,10 @@
+package org.allaymc.api.blockentity.interfaces;
+
+import org.allaymc.api.blockentity.BlockEntity;
+import org.allaymc.api.blockentity.component.BlockEntityDecoratedPotBaseComponent;
+
+/**
+ * @author IWareQ
+ */
+public interface BlockEntityDecoratedPot extends BlockEntity, BlockEntityDecoratedPotBaseComponent {
+}