+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemDecoratedPotStack.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.item.interfaces;
 
 import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.item.component.ItemDecoratedPotBaseComponent;
 
-public interface ItemDecoratedPotStack extends ItemStack {
+public interface ItemDecoratedPotStack extends ItemStack, ItemDecoratedPotBaseComponent {
 }