+++ server/src/main/java/org/allaymc/server/block/impl/BlockDecoratedPotBehaviorImpl.java
@@ -1,14 +1,19 @@
 package org.allaymc.server.block.impl;
 
+import lombok.experimental.Delegate;
+import org.allaymc.api.block.component.BlockBlockEntityHolderComponent;
 import org.allaymc.api.block.interfaces.BlockDecoratedPotBehavior;
+import org.allaymc.api.blockentity.interfaces.BlockEntityDecoratedPot;
 import org.allaymc.api.component.Component;
 import org.allaymc.server.component.ComponentProvider;
 
 import java.util.List;
 
 public class BlockDecoratedPotBehaviorImpl extends BlockBehaviorImpl implements BlockDecoratedPotBehavior {
-    public BlockDecoratedPotBehaviorImpl(
-            List<ComponentProvider<? extends Component>> componentProviders) {
+    @Delegate
+    private BlockBlockEntityHolderComponent<BlockEntityDecoratedPot> blockEntityHolderComponent;
+
+    public BlockDecoratedPotBehaviorImpl(List<ComponentProvider<? extends Component>> componentProviders) {
         super(componentProviders);
     }
 }